# Conflicts:
#	src/main/java/com/studycollaboproject/scope/service/UserService.java

+++ src/main/java/com/studycollaboproject/scope/service/TeamService.java
@@ -2,11 +2,9 @@ package com.studycollaboproject.scope.service;
 
 import com.studycollaboproject.scope.dto.MemberListResponseDto;
 import com.studycollaboproject.scope.exception.ErrorCode;
-import com.studycollaboproject.scope.exception.RestApiException;
-import com.studycollaboproject.scope.model.Applicant;
-import com.studycollaboproject.scope.model.Post;
-import com.studycollaboproject.scope.model.Team;
-import com.studycollaboproject.scope.model.User;
+import com.studycollaboproject.scope.exception.BadRequestException;
+import com.studycollaboproject.scope.exception.ForbiddenException;
+import com.studycollaboproject.scope.model.*;
 import com.studycollaboproject.scope.repository.ApplicantRepository;
 import com.studycollaboproject.scope.repository.TeamRepository;
 import lombok.RequiredArgsConstructor;
@@ -25,23 +23,26 @@ public class TeamService {
     private final ApplicantRepository applicantRepository;
 
     @Transactional
-    public Team acceptMember(Post post, User user, Boolean accept) {
+    public void acceptMember(Post post, User user, Boolean accept) {
         // [예외처리] 대기열에서 신청자의 정보를 찾을 수 없을 때
         Applicant applicant = applicantRepository.findByUserAndPost(user, post).orElseThrow(
-                () -> new RestApiException(ErrorCode.NO_APPLICANT_ERROR)
+                () -> new BadRequestException(ErrorCode.NO_APPLICANT_ERROR)
         );
+        if(accept && post.getTotalMember() == post.getRecruitmentMember()){
+            throw new BadRequestException(ErrorCode.NO_EXCEED_MEMBER_ERROR);
+        }
         //대기열에서 삭제
         applicant.deleteApply();
         applicantRepository.delete(applicant);
 
-        if(accept){
+        if (accept) {
             Team team = Team.builder()
                     .user(user)
                     .post(post)
                     .build();
-            return teamRepository.save(team);
+            teamRepository.save(team);
         }
-        return null;
+
     }
 
     @Transactional
@@ -50,11 +51,22 @@ public class TeamService {
                 .stream().map(MemberListResponseDto::new).collect(Collectors.toCollection(ArrayList::new));
     }
 
-    public void memberResignation(User member,Post post) {
-        teamRepository.deleteByUserAndPost(member,post);
+    @Transactional
+    public void memberDelete(User user, Post post) {
+        if (post.getUser().getId().equals(user.getId())) {
+            throw new BadRequestException(ErrorCode.NOT_AVAILABLE_ACCESS);
+        }
+        else if (post.getProjectStatus().equals(ProjectStatus.PROJECT_STATUS_RECRUITMENT)) {
+            Team team = teamRepository.findByUserAndPost(user, post).orElseThrow(
+                    () -> new ForbiddenException(ErrorCode.NO_TEAM_ERROR)
+            );
+            team.deleteTeam();
+            teamRepository.deleteByUserAndPost(user, post);
+        } else throw new BadRequestException(ErrorCode.NO_RECRUITMENT_ERROR);
     }
 
-    public void memberSecession(User user, Post post) {
-        teamRepository.deleteByUserAndPost(user,post);
+    public boolean isMember(Post post, User user) {
+        return teamRepository.findByUserAndPost(user,post).isPresent();
+
     }
 }
